Feat: 로그인을 안할 시 제보 및 채팅방 접속 불가 메세지 Snackbar 처리 수정
